+++ src/test/java/kr/or/ddit/test/config/WebTestConfig.java
@@ -12,7 +12,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
-
 // 20210311 KWS
 @ContextConfiguration(locations = { "classpath:/kr/or/ddit/config/spring/datasource-context.xml",
 		"classpath:/kr/or/ddit/config/spring/application-context.xml",
@@ -35,6 +34,5 @@ public class WebTestConfig {
 	@Ignore
 	@Test
 	public void dummy() {
-
 	}
 }
